Initialize dockerfile, update styling, add circuit components for display

+++ client/src/main/scala/geotrellis/admin/client/facades/LeafletFacade.scala
@@ -1505,7 +1505,7 @@ class LMap(id: String, options: UndefOr[LMapOptions]) extends LEventEmitter {
 
   def getCenter(): LLatLng = js.native
 
-  def getZoom(): Double = js.native
+  def getZoom(): Int = js.native
 
   def getMinZoom(): Double = js.native
